requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-rest-multiple-models==2.1.3
djangorestframework==3.11.2
idna==3.2
pycodestyle==2.7.0
pytz==2021.3
PyYAML==5.3.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
